🔨 refactor: util 함수 분리 및 convert관련 고차함수 변환
