+++ src/hooks/useImagePreSetup.ts
@@ -1,4 +1,5 @@
 import { useEffect, useState } from 'react';
+import { fileImgToWebP } from '../utils';
 
 interface UseImagePreSetupProps {
   imageFiles: File[] | null;
@@ -19,86 +20,17 @@ interface ProcessedFileResult {
   previewUrl: string | null;
 }
 
-const imageToWebP = async (file: File, quality: number): Promise<Blob> => {
-  return new Promise((resolve, reject) => {
-    const img = new Image();
-    const url = URL.createObjectURL(file);
-
-    img.onload = () => {
-      URL.revokeObjectURL(url);
-      const canvas = document.createElement('canvas');
-      const context = canvas.getContext('2d');
-
-      if (!context) {
-        reject(new Error('Canvas Context가 존재하지 않습니다'));
-        return;
-      }
-
-      canvas.width = img.width;
-      canvas.height = img.height;
-      context.drawImage(img, 0, 0);
-      canvas.toBlob(
-        (blob) => {
-          blob
-            ? resolve(blob)
-            : reject(new Error('이미지를 WebP형태로 변환하는데 실패했습니다.'));
-        },
-        'image/webp',
-        quality
-      );
-    };
-
-    img.onerror = () => {
-      URL.revokeObjectURL(url);
-      reject(
-        new Error(
-          '이미지 로드 중 오류가 발생했습니다. 이미지를 불러올 수 없습니다.'
-        )
-      );
-    };
-
-    img.src = url;
-  });
-};
-
 const convertFile = async (
   file: File,
-  convertToWebP: boolean,
-  quality: number
+  convertHandler: (file: File) => Promise<ProcessedFileResult>,
+  idx: number
 ): Promise<ProcessedFileResult> => {
   try {
-    if (convertToWebP) {
-      const webpBlob = await imageToWebP(file, quality);
-      const webpUrl = URL.createObjectURL(webpBlob);
-      return { webpBlob, previewUrl: webpUrl };
-    } else {
-      const url = URL.createObjectURL(file);
-      return { webpBlob: null, previewUrl: url };
-    }
+    return await convertHandler(file);
   } catch (error) {
-    console.error('파일 처리 중 오류 발생:', error);
+    console.error(`${idx + 1}번째 파일 처리 중 오류 발생:`, error);
     return { webpBlob: null, previewUrl: null };
   }
-};
-
-const getProcessResults = (
-  results: PromiseSettledResult<ProcessedFileResult>[]
-) => {
-  const previewUrlsList: (string | null)[] = [];
-  const webpImagesList: (Blob | null)[] = [];
-
-  results.forEach((result) => {
-    if (result.status === 'fulfilled') {
-      const { webpBlob, previewUrl } = result.value;
-      webpImagesList.push(webpBlob);
-      previewUrlsList.push(previewUrl);
-    } else {
-      webpImagesList.push(null);
-      previewUrlsList.push(null);
-    }
-  });
-
-  return { previewUrlsList, webpImagesList };
 };
 
 const validateWebPQuality = (convertToWebP: boolean, webPQuality: number) => {
@@ -125,38 +57,57 @@ const useImagePreSetup = ({
   convertToWebP = false,
   webPQuality = DEFAULT_WEBP_QUALITY,
 }: UseImagePreSetupProps): UseImagePreSetupReturns => {
-  webPQuality = validateWebPQuality(convertToWebP, webPQuality);
+  const validatedWebPQuality = validateWebPQuality(convertToWebP, webPQuality);
 
   const [isLoading, setIsLoading] = useState(false);
   const [isError, setIsError] = useState(false);
   const [previewUrls, setPreviewUrls] = useState<(string | null)[]>([]);
   const [webpImages, setWebpImages] = useState<(Blob | null)[]>([]);
 
+  const convertToWebPHandler = async (
+    file: File
+  ): Promise<ProcessedFileResult> => {
+    const webpBlob = await fileImgToWebP(file, validatedWebPQuality);
+    const webpUrl = URL.createObjectURL(webpBlob);
+    return { webpBlob, previewUrl: webpUrl };
+  };
+
+  const urlFromFileHandler = async (
+    file: File
+  ): Promise<ProcessedFileResult> => {
+    const url = URL.createObjectURL(file);
+    return { webpBlob: null, previewUrl: url };
+  };
+
+  const processImages = async (imageFiles: File[]) => {
+    setIsLoading(true);
+    setIsError(false);
+
+    try {
+      const fileHandler = convertToWebP
+        ? convertToWebPHandler
+        : urlFromFileHandler;
+
+      const convertResults = await Promise.all(
+        imageFiles.map((file, idx) => convertFile(file, fileHandler, idx))
+      );
+
+      const previewUrlsList = convertResults.map((result) => result.previewUrl);
+      const webpImagesList = convertResults.map((result) => result.webpBlob);
+
+      setPreviewUrls(previewUrlsList);
+      setWebpImages(webpImagesList);
+    } catch (error) {
+      setIsError(true);
+    } finally {
+      setIsLoading(false);
+    }
+  };
+
   useEffect(() => {
     if (!imageFiles || imageFiles.length === 0) return;
 
-    const processImages = async () => {
-      setIsLoading(true);
-      setIsError(false);
-
-      try {
-        const results = await Promise.allSettled(
-          imageFiles.map((file) =>
-            convertFile(file, convertToWebP, webPQuality)
-          )
-        );
-        const { previewUrlsList, webpImagesList } = getProcessResults(results);
-
-        setPreviewUrls(previewUrlsList);
-        setWebpImages(webpImagesList);
-      } catch (error) {
-        setIsError(true);
-      } finally {
-        setIsLoading(false);
-      }
-    };
-
-    processImages();
+    processImages(imageFiles);
 
     return () => {
       previewUrls.forEach((url) => {
