+++ src/utils/index.ts
@@ -4,3 +4,4 @@ export { isClient, hasNavigator } from './isClient';
 export { isTouchDevice } from './isTouchDevice';
 export { validators, MatchError } from './validators';
 export { imgToBlob } from './imgToBlob';
+export { fileImgToWebP } from './fileImgToWebP';
